\name{array.addVar}
\alias{array.addVar}
\title{Add variable to an array}
\usage{
array.addVar(a, var)
}
\arguments{
  \item{a}{An array}

  \item{var}{Variable needs to added to the array}
}
\value{
An array included the variable added
}
\description{
Add variable to an array
}
\details{
Add variable to an array
}
\author{
Weilin Lin
}

