\name{array.compareSDIV}
\alias{array.compareSDIV}
\title{Compare two arrays and plot the correlation between them}
\usage{
array.compareSDIV(arr.o, arr.n)
}
\arguments{
  \item{arr.o}{Old arrays}

  \item{arr.n}{New arrays}
}
\value{
correlation plot between arr.o and arr.n
}
\description{
Compare tow arrays
}
\details{
Compare two arrays and plot the correlation between them
}
\author{
Weilin Lin
}

