\name{array.dbind}
\alias{array.dbind}
\title{Change the dimnames and dimensions of an array}
\usage{
array.dbind(..., overwrite = FALSE)
}
\arguments{
  \item{...}{other parametres needed}

  \item{overwrite}{Logical}
}
\value{
An array with an new dimnames and dimensions
}
\description{
Change the dimnames and dimensions of an array
}
\details{
Change the dimnames and dimensions of an array
}
\author{
Weilin Lin
}

