\name{array.dbindList}
\alias{array.dbindList}
\title{Change the dimnames and dimensions of an array}
\usage{
array.dbindList(L, overwrite = FALSE)
}
\arguments{
  \item{L}{An list}

  \item{overwrite}{Logic}
}
\value{
An array with new dimnames and dimesions
}
\description{
Change the dimnames and dimensions of an array
}
\details{
Change the dimnames and dimensions of an array
}
\author{
Weilin Lin
}

