\name{backtest.accounting}
\alias{backtest.accounting}
\title{backtest.accounting}
\usage{
  backtest.accounting(hldDf, dataDf, cfgLst)
}
\arguments{
  \item{hldDf}{hldDf}

  \item{dataDf}{dataDf}

  \item{cfgLst}{cfg list}
}
\value{
  data frame
}
\description{
  backtest.accounting
}
\details{
  backtest.accounting
}
\author{
  Weilin Lin
}

