\name{backtest.dailySummary}
\alias{backtest.dailySummary}
\title{daily summary for backtest result}
\usage{
  backtest.dailySummary(hldDf)
}
\arguments{
  \item{hldDf}{hldDf}
}
\value{
  data frame
}
\description{
  daily summary for backtest result
}
\details{
  daily summary for backtest result
}
\author{
  Weilin Lin
}

