\name{backtest.loadAlpha}
\alias{backtest.loadAlpha}
\title{load alpha data for backtesting}
\usage{
  backtest.loadAlpha(startDate = NULL, endDate = NULL,
    cfgLst)
}
\arguments{
  \item{startDate}{startDate for backtesting}

  \item{endDate}{endDate for backtesting}

  \item{cfgLst}{configuration file for backtesting}
}
\value{
  cfgLst
}
\description{
  load alpha data for backtesting
}
\details{
  load alpha data for backtesting
}
\author{
  Weilin Lin
}

