\name{backtest.simpleRebalance}
\alias{backtest.simpleRebalance}
\title{simple rebalance function for backtesting}
\usage{
  backtest.simpleRebalance(hldDf, alphaDf, dataDf, univVec,
    cfgLst)
}
\arguments{
  \item{hldDf}{holding data frame}

  \item{alphaDf}{alpha data frame}

  \item{dataDf}{data frame}

  \item{univVec}{univ vector}

  \item{cfgLst}{configuration}
}
\value{
  cfgLst
}
\description{
  simple rebalance function for backtesting
}
\details{
  simple rebalance function for backtesting
}
\author{
  Weilin Lin
}

