\name{CF}
\alias{CF}
\title{Calculate cash flow of a coupon}
\usage{
CF(yrMatVec, cVec, region)
}
\arguments{
  \item{yrMatVec}{years to maturity}

  \item{cVec}{coupon}

  \item{region}{region of coupon}
}
\value{
Cash flow
}
\description{
Calculate cash flow of a coupon
}
\details{
Calculate cash flow of a coupon
}
\author{
Weilin Lin
}

