\name{colStdevs}
\alias{colStdevs}
\title{Calculate the sd of a matrix by column}
\usage{
colStdevs(x, na.rm = F)
}
\arguments{
  \item{x}{Matix}

  \item{na.rm}{Logical}
}
\value{
An vector contains the sds.
}
\description{
Calculate the sd of a matrix by column
}
\details{
Calculate the sd of a matrix by column
}
\author{
Weilin Lin
}

