\name{date.busDiff}
\alias{date.busDiff}
\title{Diff between date and BusDate}
\usage{
date.busDiff(fromDate, toDate, region)
}
\arguments{
  \item{fromDate}{start Date}

  \item{toDate}{end Date}

  \item{region}{region of stocks or bonds}
}
\value{
out Diff between date and busDate
}
\description{
Unconfirmed
}
\details{
Unconfirmed
}
\author{
Weilin Lin
}

