\name{date.checkStartEndDate}
\alias{date.checkStartEndDate}
\title{Check start and end date}
\usage{
date.checkStartEndDate(startDate, endDate, region)
}
\arguments{
  \item{startDate}{start Date}

  \item{endDate}{End Date}

  \item{region}{region}
}
\value{
StartDate and endDate
}
\description{
Check start and end date
}
\details{
Check start and end date
}
\author{
Weilin Lin
}

