\name{date.convert}
\alias{date.convert}
\title{Use dot.date.convert to Convert time format}
\usage{
date.convert(time, format = NULL)
}
\arguments{
  \item{time}{Time to convert to new format}

  \item{format}{New format of Time}
}
\value{
An object with new format
}
\description{
Use dot.date.convert to Convert time format
}
\details{
Use dot.date.convert to Convert time format
}
\author{
Weilin Lin
}

