\name{date.currentDate}
\alias{date.currentDate}
\title{Gets the current time and date, optionally for the specified zone or region.}
\usage{
date.currentDate()
}
\value{
date in format of "YYYYMMDD"
}
\description{
Gets the current time and date, optionally for the
specified zone or region.
}
\details{
Gets the current time and date, optionally for the
specified zone or region.
}
\author{
Weilin Lin
}

