\name{date.date}
\alias{date.date}
\title{Convert date to format "YYYYMMDD"}
\usage{
date.date(time)
}
\arguments{
  \item{time}{time to be converted}
}
\value{
time converted
}
\description{
Convert date to format "YYYYMMDD"
}
\details{
Convert date to format "YYYYMMDD"
}
\author{
Weilin Lin
}

