\name{date.diffMonth}
\alias{date.diffMonth}
\title{Calculate the diff between two monthes}
\usage{
date.diffMonth(m1, m2, inFormat = "YYYYMM")
}
\arguments{
  \item{m1}{any month}

  \item{m2}{any month}

  \item{inFormat}{the format of diff}
}
\value{
diff
}
\description{
Calculate the diff between two monthes
}
\details{
Calculate the diff between two monthes
}
\author{
Weilin Lin
}

