\name{date.fromJulian}
\alias{date.fromJulian}
\title{Caculates dates bases date.R.ORIGIN}
\usage{
date.fromJulian(julian)
}
\arguments{
  \item{julian}{Julian time (days since some origin)}
}
\value{
a date
}
\description{
Caculates dates bases date.R.ORIGIN
}
\details{
Caculates dates bases date.R.ORIGIN
}
\examples{
date.fromJulian(100)
}
\author{
Weilin Lin
}

