\name{date.isBusDate}
\alias{date.isBusDate}
\title{check if the day is trading day}
\usage{
date.isBusDate(date, region = "CN")
}
\arguments{
  \item{date}{date in YYYYMMDD format}

  \item{region}{specify the market, CN or HK are supported}
}
\value{
TRUE or FALSE
}
\description{
check if the day is trading day
}
\details{
check if the day is trading day
}
\author{
Weilin Lin
}

