\name{date.month}
\alias{date.month}
\title{Extract the month of date}
\usage{
date.month(date)
}
\arguments{
  \item{date}{any date}
}
\value{
the year and month as numeric
}
\description{
Extract the month of date
}
\details{
Extract the month of date
}
\author{
Weilin Lin
}

