\name{date.monthToFirstBusDate}
\alias{date.monthToFirstBusDate}
\title{month to first Bus Date}
\usage{
date.monthToFirstBusDate(month, region)
}
\arguments{
  \item{month}{month}

  \item{region}{region}
}
\value{
null
}
\description{
month to first Bus Date
}
\details{
month to first Bus Date
}
\author{
Weilin Lin
}

