\name{date.monthToLastBusDate}
\alias{date.monthToLastBusDate}
\title{month to last BusDate}
\usage{
date.monthToLastBusDate(month, region)
}
\arguments{
  \item{month}{month}

  \item{region}{region for month}
}
\value{
null
}
\description{
month to last BusDate
}
\details{
month to last BusDate
}
\author{
Weilin Lin
}

