\name{date.range}
\alias{date.range}
\title{date range}
\usage{
date.range(startDate, endDate = NULL, n = NULL, region)
}
\arguments{
  \item{startDate}{start Date of date range}

  \item{endDate}{end Date for date range}

  \item{n}{dyes between startDate and endDate}

  \item{region}{region for stocks or bonds}
}
\value{
date range
}
\description{
date range
}
\details{
date range
}
\author{
Weilin Lin Weilin Lin
}

