\name{date.secondsSinceMidnight}
\alias{date.secondsSinceMidnight}
\title{seconds from midnight}
\usage{
date.secondsSinceMidnight(timeOrTimeString)
}
\arguments{
  \item{timeOrTimeString}{}
}
\value{
date string
}
\description{
seconds from midnight
}
\details{
seconds from midnight
}
\author{
Weilin Lin
}

