\name{date.shiftMonths}
\alias{date.shiftMonths}
\title{shift months}
\usage{
date.shiftMonths(months, shift)
}
\arguments{
  \item{months}{original months}

  \item{shift}{shift of months}
}
\value{
a months shifted
}
\description{
shift months
}
\details{
shift months
}
\author{
Weilin Lin
}

