\name{date.toBusDate}
\alias{date.toBusDate}
\title{Convert date to BusDate}
\usage{
date.toBusDate(date, region, shift = 0, forward = TRUE)
}
\arguments{
  \item{date}{date to be converted}

  \item{region}{region for stocks or bonds}

  \item{shift}{0}

  \item{forward}{logical}
}
\value{
DATE.BUS.DATES
}
\description{
Convert date to BusDate
}
\details{
Convert date to BusDate
}
\author{
Weilin Lin
}

