\name{date.toString}
\alias{date.toString}
\title{convert date to strings}
\usage{
date.toString(time)
}
\arguments{
  \item{time}{time to be converted}
}
\value{
time converted
}
\description{
convert date to strings
}
\details{
convert date to strings
}
\author{
Weilin Lin
}

