\name{date2num}
\alias{date2num}
\title{transfer date to number}
\usage{
date2num(x)
}
\arguments{
  \item{x}{Datetime}
}
\value{
An number
}
\description{
transfer date to number
}
\details{
transfer date to number
}
\examples{
date2num("2013-01-01")
}
\author{
Yishuo Deng
}

