\name{datetime2POSIXlt}
\alias{datetime2POSIXlt}
\title{transfer datetime to POSIXlt}
\usage{
datetime2POSIXlt(x)
}
\arguments{
  \item{x}{Datetime}
}
\value{
An POSIXlt object
}
\description{
transfer datetime to POSIXlt
}
\details{
transfer datetime to POSIXlt
}
\author{
Yishuo Deng
}

