\name{day_seq}
\alias{day_seq}
\title{Generate a time series}
\usage{
day_seq(startdate, enddate, ...)
}
\arguments{
  \item{startdate}{startdate of time series}

  \item{enddate}{endDate of time series}
}
\value{
xts object
}
\description{
Generate a time series
}
\details{
Generate a time series
}
\examples{
day_seq(20120302,20130405)
}
\author{
Yishuo Deng
}

