\name{des_date}
\alias{des_date}
\title{Discompose an Date to day,month and year}
\usage{
des_date(x, ...)
}
\arguments{
  \item{x}{The Date object needed to be discomposed}

  \item{...}{Other parameters}
}
\value{
An dataframe contains the day, month and year
}
\description{
Discompose an Date to day,month and year
}
\details{
Discompose an Date to day,month and year
}
\examples{
des_date(as.Date("2013-03-04"))
}
\author{
Yishuo Deng
}

