\name{df.arrayToDf}
\alias{df.arrayToDf}
\title{Transfer an array to Data frame}
\usage{
df.arrayToDf(a, valueCol = "value", valueDim = NULL)
}
\arguments{
  \item{a}{an arry}

  \item{valueCol}{Column contains the value}

  \item{valueDim}{logical}
}
\value{
d
}
\description{
Transfer an array to Data frame
}
\details{
Transfer an array to Data frame
}
\author{
Weilin Lin
}

