\name{df.dfToArray}
\alias{df.dfToArray}
\title{Convert df to array}
\usage{
df.dfToArray(d, keyCols, valueCols, valueColDimnameName = NULL,
  drop = FALSE)
}
\arguments{
  \item{d}{uncomfirmed}

  \item{keyCols}{uncomfirmed}

  \item{valueCols}{uncomfirmed}

  \item{valueColDimnameName}{uncomfirmed}

  \item{drop}{Logical}
}
\value{
a
}
\description{
Convert df to Array
}
\details{
Convert df to Array
}
\author{
Weilin Lin
}

