\name{df.dfToSQL}
\alias{df.dfToSQL}
\title{data frame to SQL insert query}
\usage{
df.dfToSQL(data, dbname)
}
\arguments{
  \item{data}{data frame}

  \item{dbname}{database to insert}
}
\value{
SQL string
}
\description{
data frame to SQL conversion
}
\details{
Convert data frame to SQL for insert
}
\author{
Weilin Lin
}

