\name{df.rbindJson}
\alias{df.rbindJson}
\title{data frame to json conversion}
\usage{
df.rbindJson(data)
}
\arguments{
  \item{data}{data frame with jtid and time in first and
  second column}
}
\value{
json string
}
\description{
convert data frame with jtid and time to json
}
\details{
convert data frame with jtid and time to json
}
\author{
Weilin Lin
}

