\name{df.renameVar}
\alias{df.renameVar}
\title{rename variables in a data frame}
\usage{
df.renameVar(df, oldNames, newNames, warning = FALSE)
}
\arguments{
  \item{df}{data frame}

  \item{oldNames}{old names}

  \item{newNames}{new names}

  \item{warning}{warning}
}
\value{
df
}
\description{
rename variables in a data frame
}
\details{
rename variables in a data frame
}
\author{
Weilin Lin
}

