\name{df.sdivToDf}
\alias{df.sdivToDf}
\title{transforms SDIV(stock,date,invernal,vale) to data frame}
\usage{
df.sdivToDf(arr, verbose = FALSE)
}
\arguments{
  \item{arr}{array}

  \item{verbose}{logical}
}
\value{
arr
}
\description{
transforms SDIV(stock,date,invernal,vale) to data frame
}
\details{
transforms SDIV(stock,date,invernal,vale) to data frame
}
\author{
Weilin Lin
}

