\name{df.summary}
\alias{df.summary}
\title{Calculates the summary information of a data frame}
\usage{
df.summary(df)
}
\arguments{
  \item{df}{data frame}
}
\value{
xx
}
\description{
Calculates the summary information of a data frame
}
\details{
Calculates the summary information of a data frame
}
\author{
Weilin Lin
}

