\name{df.weighted.mean}
\alias{df.weighted.mean}
\title{calculates weighted mean of data frame}
\usage{
df.weighted.mean(df, meanVar, byVar, wgtVar, ...)
}
\arguments{
  \item{df}{data frame}

  \item{meanVar}{variables to calculates mean}

  \item{byVar}{group variables}

  \item{wgtVar}{weight variables}

  \item{...}{other parameters}
}
\value{
data
}
\description{
calculates weighted mean of data frame
}
\details{
calculates weighted mean of data frame
}
\author{
Weilin Lin
}

