\name{dot.cf}
\alias{dot.cf}
\title{Calculate cash flow of a coupon}
\usage{
dot.cf(yrMat, c, region = "CN")
}
\arguments{
  \item{yrMat}{years to maturity}

  \item{c}{coupon}

  \item{region}{region of coupon}
}
\value{
cash flow of the coupon
}
\description{
Calculate cash flow of a coupon
}
\details{
Calculate cash flow of a coupon
}
\author{
Weilin Lin
}

