\name{dot.date.convert}
\alias{dot.date.convert}
\title{Convert date format to POSIXct}
\usage{
dot.date.convert(time, format = "\%Y\%m\%d")
}
\arguments{
  \item{time}{time to be converted}

  \item{format}{Time format}
}
\value{
An formated time
}
\description{
convert date format to POSIXct
}
\details{
Convert date format to POSIXct
}
\author{
Weilin Lin
}

