\name{irr2reg}
\alias{irr2reg}
\title{Transfer irregular time series to regular}
\usage{
irr2reg(x, ...)
}
\arguments{
  \item{x}{An irregular time series}

  \item{...}{Other parameters}
}
\value{
An regular time series
}
\description{
Transfer irregular time series to regular
}
\details{
Transfer irregular time series to regular
}
\author{
Yishuo Deng
}

