\name{kurtosis}
\alias{kurtosis}
\title{calculates kurtosis}
\usage{
kurtosis(x, na.rm = F, method = "fisher")
}
\arguments{
  \item{x}{vector}

  \item{na.rm}{logical}

  \item{method}{method used to calculates kurtosis}
}
\value{
kurtosis
}
\description{
calculates kurtosis
}
\details{
calculates kurtosis
}
\author{
Weilin Lin
}

