\name{match.data.frame}
\alias{match.data.frame}
\title{match two lists}
\usage{
match.data.frame(x, table, nomatch = NA, incomparables = F)
}
\arguments{
  \item{x}{list}

  \item{table}{a list the same length as x}

  \item{nomatch}{the value to be returned in the case when
  no match is found. Note that it is coerced to integer.}

  \item{incomparables}{logical}
}
\value{
A vector of the same length as x
}
\description{
match two lists
}
\details{
match two lists
}
\author{
Weilin Lin
}

