\name{plot.HFT}
\alias{plot.HFT}
\title{plot function for HFT data}
\usage{
\method{plot}{HFT}(data, var = NULL, ...)
}
\arguments{
  \item{data}{}

  \item{var}{plot var}

  \item{...}{}
}
\description{
plot function for HFT data
}
\details{
plot function for HFT data
}
\author{
Weilin Lin
}

