\name{rankofday}
\alias{rankofday}
\title{Calulates the rank of a Date}
\usage{
rankofday(x, ...)
}
\arguments{
  \item{Date}{A date.}

  \item{...}{Other parameters.}
}
\value{
The rank of the date
}
\description{
Calulates the rank of a Date
}
\details{
Calulates the rank of a Date
}
\examples{
rankofday("2013-01-25");rankofday(as.Date("2013-01-25"))
}
\author{
Yishuo Deng
}

