\name{skewness}
\alias{skewness}
\title{Calculates the skewness of a distribution}
\usage{
skewness(x, na.rm = F, method = "fisher")
}
\arguments{
  \item{x}{Data}

  \item{na.rm}{Logical}

  \item{method}{menthod usded to calculates skewness}
}
\value{
skewness
}
\description{
Calculates the skewness of a distribution
}
\details{
Calculates the skewness of a distribution
}
\author{
Weilin Lin
}

