\docType{data}
\name{tablename}
\alias{tablename}
\title{read csv data}
\format{chr "daily.mktData"}
\usage{
  tablename
}
\arguments{
  \item{startDate}{start Date for data}

  \item{endDate}{end Date for data}

  \item{intervals}{intervals for data}

  \item{shift}{shift}

  \item{jtids}{symbols for stock or bonds,must be
  charactor.}

  \item{vars}{variables in data}

  \item{region}{region for data}

  \item{...}{other parametres}
}
\value{
  a data file
}
\description{
  read csv data
}
\details{
  read csv data
}
\author{
  Yishuo Deng
}
\keyword{datasets}

