\name{utils.assertEquals}
\alias{utils.assertEquals}
\title{assert whether two objects are equal}
\usage{
utils.assertEquals(expected, generated, msg = NULL,
  printInputsOnDiff = TRUE, errorFunc = "stop", ...)
}
\arguments{
  \item{expected}{result expected}

  \item{generated}{result generated}

  \item{msg}{massage}

  \item{printInputsOnDiff}{logical}

  \item{errorFunc}{function for error}

  \item{...}{other parameters needed}
}
\value{
logical
}
\description{
assert whether two objects are equal
}
\details{
assert whether two objects are equal
}
\author{
Weilin Lin
}

