\name{utils.combineFactors}
\alias{utils.combineFactors}
\title{transforms data frame to list of character}
\usage{
utils.combineFactors(df, sep = ":")
}
\arguments{
  \item{df}{data frame}

  \item{sep}{the field separator character.}
}
\value{
an list
}
\description{
transforms data frame to list of character
}
\details{
transforms data frame to list of character
}
\author{
Weilin Lin
}

