\name{utils.intersectList}
\alias{utils.intersectList}
\title{calculates the intersect of lis}
\usage{
utils.intersectList(lst)
}
\arguments{
  \item{lst}{list}
}
\value{
x,the intersect
}
\description{
calculates the intersect of list
}
\details{
calculates the intersect of lis
}
\author{
Weilin Lin
}

