\name{utils.pad}
\alias{utils.pad}
\title{Padding string}
\usage{
utils.pad(string, lengthOut = 2, padChar = "0", leading = TRUE)
}
\arguments{
  \item{string}{Any string}

  \item{lengthOut}{lenthOut}

  \item{padChar}{Padding string}

  \item{leading}{Logical}
}
\value{
x
}
\description{
Padding string
}
\details{
Padding string
}
\author{
Weilin Lin
}

