\name{utils.pctFilter}
\alias{utils.pctFilter}
\title{filer stock return}
\usage{
utils.pctFilter(stockRet, probs = c(0.01, 0.99))
}
\arguments{
  \item{stockRet}{stock returns}

  \item{probs}{probility}
}
\value{
null
}
\description{
filer stock return
}
\details{
filer stock return
}
\author{
Weilin Lin
}

