\name{utils.readCSVBus}
\alias{utils.readCSVBus}
\title{utils.readCSVBus}
\usage{
utils.readCSVBus(jtids = NULL, startDate, endDate, dataDir,
  filePrefix = "data", intervals = NULL, vars = NULL, shift = 0,
  region = NULL, verbose = FALSE, outClass = "data.frame",
  fileNameDelimiter = "_", ...)
}
\arguments{
  \item{jtids}{jtids}

  \item{startDate}{startDate}

  \item{endDate}{endDate}

  \item{dataDir}{dataDir}

  \item{filePrefix}{filePrefix}

  \item{intervals}{intervals}

  \item{vars}{vars}

  \item{shift}{shift}

  \item{region}{region}

  \item{verbose}{verbose}

  \item{outClass}{outClass}

  \item{fileNameDelimiter}{fileNameDelimiter}

  \item{...}{}
}
\value{
data frame
}
\description{
utils.readCSVBus
}
\details{
utils.readCSVBus
}
\author{
Weilin Lin
}

