\name{utils.readMySQL}
\alias{utils.readMySQL}
\title{read csv data}
\usage{
utils.readMySQL(tablename = NULL, startDate, endDate, Dates = NULL,
  jtids = NULL, vars = NULL, region = NULL, ...)
}
\arguments{
  \item{startDate}{start Date for data}

  \item{endDate}{end Date for data}

  \item{intervals}{intervals for data}

  \item{shift}{shift}

  \item{jtids}{symbols for stock or bonds,must be
  charactor.}

  \item{vars}{variables in data}

  \item{region}{region for data}

  \item{...}{other parametres}
}
\value{
a data file
}
\description{
read csv data
}
\details{
read csv data
}
\author{
Yishuo Deng
}

