\name{utils.substring}
\alias{utils.substring}
\title{extract substrings of text}
\usage{
utils.substring(text, first, last = NULL)
}
\arguments{
  \item{text}{a character vector.}

  \item{first}{integer. The first element to be replaced.}

  \item{last}{integer. The last element to be replaced.}
}
\value{
substring is compatible with S, with first and last instead
of start and stop. For vector arguments, it expands the
arguments cyclically to the length of the longest provided
none are of zero length.
}
\description{
extract substrings of text
}
\details{
extract substrings of text
}
\author{
Weilin Lin
}

