\name{utils.summary}
\alias{utils.summary}
\title{Calculates the summary information of an vector}
\usage{
utils.summary(x)
}
\arguments{
  \item{x}{Any objects}
}
\value{
The Stdev,Skewness and kurtosis of the data
}
\description{
Calculates the summary information of an vector
}
\details{
Calculates the summary information of an vector
}
\author{
Weilin Lin
}

