% Generated by roxygen2 (4.0.1): do not edit by hand
\name{which.na}
\alias{which.na}
\title{Find the index of NAs in a vector}
\usage{
which.na(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The index of NAs in the vector
}
\description{
Find the index of NAs in a vector
}
\details{
Find the index of NAs in a vector
}
\author{
Weilin Lin Weilin Lin
}

